% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_blogdown_helpers.R
\name{addin_details}
\alias{addin_details}
\title{Insert expandable details block}
\usage{
addin_details()
}
\description{
Call this function as an addin to insert at the cursor position an expandable
HTML details block.
}
