% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_blogdown_helpers.R
\name{addin_img_accessible}
\alias{addin_img_accessible}
\title{Insert accessible image HTML}
\usage{
addin_img_accessible()
}
\description{
Call this function as an addin to insert at the cursor position an HTML block
that allows for the alt text and caption to differ, unlike \code{ ![]() }
syntax.
}
