% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_blogdown_helpers.R
\name{addin_new_tab}
\alias{addin_new_tab}
\title{Insert link that opens in a new tab}
\usage{
addin_new_tab()
}
\description{
Call this function as an addin to insert at the cursor position a web link
that opens in a new tab rather than the current tab.
}
