% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinwip-package.R
\docType{package}
\name{addinwip-package}
\alias{addinwip}
\alias{addinwip-package}
\title{addinwip: Insert Useful Code Snippets For Blogdown Posts.}
\description{
RStudio addins to insert code snippets that make my life a bit
    easier to generate.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
