% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-helpers.R
\name{bs_add_biscuits}
\alias{bs_add_biscuits}
\title{Add Closing Paren Labels [WIP]}
\usage{
bs_add_biscuits()
}
\description{
In development. Adds comments after closing parentheses in your selection 
with the name of the function being closed (sometimes called 'biscuits'). 
Warning: also styles the file. Please check the output.
}
