% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-helpers.R
\name{bs_blank_target}
\alias{bs_blank_target}
\title{Add Blank Target to Links}
\usage{
bs_blank_target()
}
\description{
In selected text, append \code{{target='_blank'}} to all links that are in
the form \code{[]()}.
}
