% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_details_expand}
\alias{bs_details_expand}
\title{Insert Expandable Details Block}
\usage{
bs_details_expand()
}
\description{
Call this function as an addin to insert at the cursor position an expandable
HTML details block.
}
