% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-helpers.R
\name{bs_embed_comment}
\alias{bs_embed_comment}
\title{Insert knitr comment engine}
\usage{
bs_embed_comment()
}
\description{
Modify the selected text into the knitr comment engine.
As described \href{https://github.com/rstudio/rmarkdown/issues/1619#issuecomment-1010045812}{here}.
}
