% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_embed_tweet}
\alias{bs_embed_tweet}
\title{Embed Tweet}
\usage{
bs_embed_tweet()
}
\description{
Replace a Twitter shortcode with a chunk that contains code to embed the
tweet. Also names the chunk after the shortcode.
}
