% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_img_accessible}
\alias{bs_img_accessible}
\title{Insert Accessible Image HTML}
\usage{
bs_img_accessible()
}
\description{
Call this function as an addin to insert at the cursor position an HTML block
that allows for the alt text and caption to differ, unlike \code{ ![]() }
syntax.
}
