% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_insert_link}
\alias{bs_insert_link}
\title{Insert Link That Opens in a New Tab}
\usage{
bs_insert_link()
}
\description{
Call this function as an addin to insert at the cursor position some Markdown
that will open a web link in a new tab rather than the current tab.
}
