% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_make_link}
\alias{bs_make_link}
\title{Make Into Link}
\usage{
bs_make_link()
}
\description{
Modify the selected text or URL into the form \code{[](){target='_blank'}}.
Thanks to
\href{https://www.hvitfeldt.me/blog/creating-rstudio-addin-to-modify-selection/}{Emil
Hvitfeldt's blog post} for guidance.
}
