% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-helpers.R
\name{bs_named_anchor}
\alias{bs_named_anchor}
\title{Add Named Anchor}
\usage{
bs_named_anchor()
}
\description{
Add a named anchor to the selected text, generating a hyphenated tag in the
process. Remember to add '{#anchor-tag}' to where your anchoring to.
}
