% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_replace_link}
\alias{bs_replace_link}
\title{Add to Selection a Link to Open in a New Tab}
\usage{
bs_replace_link()
}
\description{
Call this function as an addin to replace selected text with Markdown to an
external link that opens in a new tab rather than the current tab. Thanks to
\href{https://www.hvitfeldt.me/blog/creating-rstudio-addin-to-modify-selection/}{Emil
Hvitfeldt's blog post} for guidance.
}
