% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-helpers.R
\name{bs_session_details}
\alias{bs_session_details}
\title{Insert Session Details}
\usage{
bs_session_details()
}
\description{
Inserts at the cursor position an expandable HTML details block that contains
the date of the last update and session information.
}
