% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_target_blank}
\alias{bs_target_blank}
\title{Add {target='_blank'}}
\usage{
bs_target_blank()
}
\description{
In selected text, append \code{{target='_blank'}} to links that are in
the form \code{[]()}.
}
