% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-package.R
\docType{package}
\name{blogsnip-package}
\alias{blogsnip}
\alias{blogsnip-package}
\title{blogsnip: Misc RStudio Addins To Insert Code Snippets}
\description{
Some RStudio addins for personal use that insert code snippets to
    help me write R Markdown documents, obstensibly {blogdown} posts.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
