% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogsnip-package.R
\docType{package}
\name{blogsnip-package}
\alias{blogsnip}
\alias{blogsnip-package}
\title{blogsnip: Insert Useful Code Snippets For Blogdown Posts}
\description{
RStudio addins to insert code snippets that are useful for
    writing Blogdown posts.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
