% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_blank_target}
\alias{bs_blank_target}
\title{Add Blank Target to Link}
\usage{
bs_blank_target()
}
\description{
In selected text, append \code{{target='_blank'}} to links that are in
the form \code{[]()}. Currently works only when the highlighted text has only
one link to replace.
}
