% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_img_accessible}
\alias{bs_img_accessible}
\title{Insert Accessible Image}
\usage{
bs_img_accessible()
}
\description{
Inserts at the cursor position some HTML image code that has caption and alt
text elements.
}
