% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosnip-helpers.R
\name{bs_session_details}
\alias{bs_session_details}
\title{Insert Expandable Session Details Block}
\usage{
bs_session_details()
}
\description{
Call this function as an addin to insert at the cursor position an expandable
HTML details block containing an R Markdown chunk containing the session info
and current system date.
}
